requirements.txt
mouse==0.7.1
Pillow==9.2.0
pystray==0.19.4
six==1.16.0
